Refactored Settings into the Window pattern. Removed longtext and made it more of an internal class.
