--- MangaManager/src/Common/loadedcomicinfo.py
@@ -1,5 +1,0 @@
-from __future__ import annotations
-
-#move-to (from settings)
-
-
